feat: (1) re-implement touch, (2) re-implement mkdir, (3) re-implement mv, (4) add mv.into, mv.over and rm
for (1), (2) and (3) I essentially abstract away the error handling so that user do not need to care about error raised by the java API. All errors will be logged anyway and boolean are return if an operation is successful meaning that the state of the hdfs has been changed. for (4) Still some work are needed here.